prs4dummies/rag-core/requirements-rag.txt
# RAG API Dependencies
# Core dependencies (inherited from indexing requirements)
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.0
tqdm>=4.64.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
einops>=0.6.0

# CPU version (default for laptop testing)
faiss-cpu>=1.7.4

# For GPU acceleration (uncomment when moving to Colab or GPU machine)
# faiss-gpu>=1.7.4

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Additional dependencies for the RAG pipeline
accelerate>=0.20.0
safetensors>=0.3.0

# Optional: For better performance on Windows
# torch-directml>=1.13.0  # Uncomment if using DirectML on Windows
